rename calculate_suitability.py with prototype_ prefix and remove _v2.py suffix from calculate_suitability_v2.py
